Fixes to MBC1
